Relevant change in preprocess2D
This is a minor change in code but with a potential impact on EDA
results. When in a variable of the calibration data set, after centring,
all values are 0, and auto-scaling is the preprocessing method selected,
a value of 1 was selected for the standard deviation. However, this may
reduce the sensibility in MSPC or EDA of that variable. Now, the minimum
standard deviation in the remaining variables is taken and divided by
two, so that the variable with 0's is specially sensible to any value in
future data.